+++ packages/sitecore-jss-nextjs/src/editing/index.ts
@@ -7,6 +7,7 @@ export {
   EditingRenderMiddlewareConfig,
   EditingMetadataPreviewData,
   isEditingMetadataPreviewData,
+  isComponentLibraryPreviewData,
 } from './editing-render-middleware';
 export {
   EditingPreviewData,
@@ -23,3 +24,8 @@ export {
   EditingConfigMiddleware,
   EditingConfigMiddlewareConfig,
 } from './editing-config-middleware';
+export {
+  RenderingType,
+  EDITING_COMPONENT_PLACEHOLDER,
+  EDITING_COMPONENT_ID,
+} from '@sitecore-jss/sitecore-jss/layout';