+++ packages/sitecore-jss-react/src/tests/jsdom-setup.ts
@@ -10,7 +10,9 @@ declare var global: NodeJS.Global;
 
 const { JSDOM } = require('jsdom');
 
-const jsdom = new JSDOM('<!doctype html><html><body></body></html>');
+const jsdom = new JSDOM('<!doctype html><html><body></body></html>', {
+  url: 'http://localhost',
+});
 const jsDomWindow = jsdom.window;
 
 /**
@@ -34,6 +36,7 @@ function copyProps(src: unknown, target: { [key: string]: unknown }) {
 global.window = jsDomWindow;
 global.document = jsDomWindow.document;
 global.navigator['#userAgent'] = 'node.js';
+global.jsdom = jsdom;
 
 global.HTMLElement = jsDomWindow.HTMLElement; // makes chai "happy" https://github.com/chaijs/chai/issues/1029
 copyProps(jsDomWindow, global);