+++ packages/sitecore-jss/src/layout/index.ts
@@ -16,6 +16,9 @@ export {
   ComponentParams,
   EditMode,
   FieldMetadata,
+  RenderingType,
+  EDITING_COMPONENT_PLACEHOLDER,
+  EDITING_COMPONENT_ID,
 } from './models';
 
 export {