+++ packages/sitecore-jss/src/layout/models.ts
@@ -9,6 +9,7 @@ export interface LayoutServiceData {
 
 /**
  * Layout Service page state enum
+ * library mode would render a single component
  */
 export enum LayoutServicePageState {
   Preview = 'preview',
@@ -39,6 +40,7 @@ export interface LayoutServiceContext {
   site?: {
     name?: string;
   };
+  renderingType?: RenderingType;
   editMode?: EditMode;
   clientScripts?: string[];
   clientData?: Record<string, Record<string, unknown>>;
@@ -159,3 +161,19 @@ export interface PlaceholderData {
   path: string;
   elements: Array<HtmlElementRendering | ComponentRendering>;
 }
+
+/**
+ * Editing rendering type
+ */
+export enum RenderingType {
+  Component = 'component',
+}
+
+/**
+ * Static placeholder name used for component rendering
+ */
+export const EDITING_COMPONENT_PLACEHOLDER = 'editing-componentmode-placeholder';
+/**
+ * Id of wrapper for component rendering
+ */
+export const EDITING_COMPONENT_ID = 'editing-component';