+++ CHANGELOG.md
@@ -11,7 +11,9 @@ Our versioning strategy is as follows:
 
 ## Unreleased
 
-## 22.4.1
+### 🎉 New Features & Improvements
+
+* `[nextjs][sitecore-jss-nextjs]` Support for Component Library feature in XMCloud ([#1987](https://github.com/Sitecore/jss/pull/1987)[#2000](https://github.com/Sitecore/jss/pull/2000)[#2002](https://github.com/Sitecore/jss/pull/2002)[#2005](https://github.com/Sitecore/jss/pull/2005))
 
 ### 🐛 Bug Fixes